requirements.txt
Django==4.2.13
django-channels==0.7.0
django-nvd3==0.10.1
djangorestframework==3.15.2
drf-api-logger==1.1.15
django-import-export[all]==4.0.1
psycopg2==2.9.9
daphne==4.1.2
sentry-sdk==2.1.1
django-admin-rangefilter==0.12.4
django-filter==24.2
django-otp==1.5.0
django-crontab==0.7.1
beautifulsoup4==4.12.3
celery==5.4.0
pillow==10.3.0
gunicorn==22.0.0
